memtx: keep track of all taken snapshots
Not only of the most recent one, as we do now. This is needed for
garbage collection, so that we could avoid scanning snap_dir to find
old snapshots. Also, introduce recovery_get_checkpoint() helper, which
returns vclock of the snapshot taken the specified number of snapshots
before the last snapshot. It will be used to respect snapshot_count
config option while collecting garbage.