vinyl: use dedicated thread for writing metadata log
We can't write to xlog from a coeio task, because coeio uses a pool of
workers while xlog can only be used from a single thread due to obuf.
As a result, we are bound to use various hacks, like reopening xlog on
rotation or encoding log records in tx. The hacks are getting more and
more ugly, nearly unbearable, especially when it comes to separating
garbage collection from log rotation. So let's introduce a separate
thread for writing vinyl metadata log.