Consolidate garbage collection
Currently, old snapshots and xlogs are deleted by the snapshot daemon
while vinyl files are removed by the Engine::waitCheckpoint callback.
As a result, snapshot_count does not control lifetime of vinyl data,
and it is impossible for tarantool core to temporarily disable garbage
collection, e.g. for the sake of replication. To fix that, let's
consolidate garbage collection code in one place - call it right after
successful snapshot. To remove engine-specific files, introduce a new
callback, Engine::collectGarbage, which is supposed to delete all files
that became unused before a given LSN.